style: divide stream_face.py into face.py, udp_face.py and stream_face.py
